Making different shard for test with and without bioptim
